refactor: Square 패키지 생성
