+++ src/main/java/chess/domain/BoardFactory.java
@@ -9,6 +9,9 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
+import chess.domain.square.File;
+import chess.domain.square.Rank;
+import chess.domain.square.Square;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
