+++ src/main/java/chess/domain/square/Square.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.square;
 
 import java.util.Objects;
